+++ packages/taro-weapp/src/util.js
@@ -118,7 +118,7 @@ function diffArrToPath (to, from, res = {}, keyPrev = '') {
             diffArrToPath(toItem, fromItem, res, `${targetKey}`)
           }
         } else {
-          if (!toItem || !fromItem) {
+          if (!toItem || !fromItem || keyList(toItem).length < keyList(fromItem).length) {
             res[targetKey] = toItem
           } else {
             // 对象
@@ -166,7 +166,7 @@ export function diffObjToPath (to, from, res = {}, keyPrev = '') {
           }
         } else {
           // null
-          if (!toItem || !fromItem) {
+          if (!toItem || !fromItem || keyList(toItem).length < keyList(fromItem).length) {
             res[targetKey] = toItem
           } else {
           // 对象
@@ -177,4 +177,41 @@ export function diffObjToPath (to, from, res = {}, keyPrev = '') {
     }
   }
   return res
+}
+
+export function queryToJson (str) {
+  const dec = decodeURIComponent 
+  const qp = str.split('&')
+  let ret = {}
+  let name
+  let val
+  for (let i = 0, l = qp.length, item; i < l; ++i) {
+    item = qp[i]
+    if (item.length) {
+      const s = item.indexOf('=')
+      if (s < 0) {
+        name = dec(item)
+        val = ''
+      } else {
+        name = dec(item.slice(0, s))
+        val = dec(item.slice(s + 1))
+      }
+      if (typeof ret[name] === 'string') { // inline'd type check
+        ret[name] = [ret[name]]
+      }
+
+      if (isArray(ret[name])) {
+        ret[name].push(val)
+      } else {
+        ret[name] = val
+      }
+    }
+  }
+  return ret // Object
+}
+
+const _loadTime = (new Date()).getTime().toString()
+let _i = 1
+export function getUniqueKey () {
+  return _loadTime + (_i++)
 }